requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
black==22.6.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.15
django-cors-headers==3.13.0
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
executing==0.10.0
Faker==12.0.1
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
ipython==8.4.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mixer==7.2.2
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.8.1
